enums moved to dto/enums
