+++ conveyor/src/main/java/com/aryunin/conveyor/dto/enums/EmploymentStatus.java
@@ -1,4 +1,4 @@
-package com.aryunin.conveyor.entity.enums;
+package com.aryunin.conveyor.dto.enums;
 
 public enum EmploymentStatus {
     EMPLOYED,
