+++ conveyor/src/main/java/com/aryunin/conveyor/dto/enums/MaterialStatus.java
@@ -1,4 +1,4 @@
-package com.aryunin.conveyor.entity.enums;
+package com.aryunin.conveyor.dto.enums;
 
 public enum MaterialStatus {
     POOR,
