+++ conveyor/src/main/java/com/aryunin/conveyor/dto/enums/Position.java
@@ -1,4 +1,4 @@
-package com.aryunin.conveyor.entity.enums;
+package com.aryunin.conveyor.dto.enums;
 
 public enum Position {
     WORKER,
